day 5 cleanup
