--- Day5/pt1/sample2.txt
@@ -1,23 +1,0 @@
-seeds: 4106085912 135215567 529248892 159537194 1281459911 114322341 1857095529 814584370 2999858074 50388481 3362084117 37744902 3471634344 240133599 3737494864 346615684 1585884643 142273098 917169654 286257440
-
-seed-to-soil map:
-1640984363 3136305987 77225710
-3469528922 1857474741 56096642
-278465165 2901870617 105516220
-1442950910 1913571383 198033453
-463085535 1458252975 13696838
-1718210073 1686050230 171424511
-383981385 3064707638 71598349
-1267048154 3759577328 175902756
-3262934306 1479455614 206594616
-2493001016 200414015 157177749
-3885112776 411057950 157348766
-4042461542 1181162257 199039568
-476782373 2111604836 790265781
-455579734 1471949813 7505801
-1889634584 3213531697 546045631
-4241501110 357591764 53466186
-3525625564 3935480084 359487212
-2650178765 568406716 612755541
-200414015 1380201825 78051150
-2435680215 3007386837 57320801
